+++ CHANGELOG.md
@@ -5,6 +5,10 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 **Changes to JS assets (including the front-end and JS libraries) are not shown here**, but in [`/browser/CHANGELOG`](/browser/CHANGELOG.md).
 See [STATUS.md](server/STATUS.md) to learn more about which features will remain stable.
 
+## [UNRELEASED]
+
+- Remove `process-management` feature #324 #334
+
 ## [v0.37.0] - 2024-02-01
 
 - Refactor `atomic_lib::Resource` propval methods (e.g. `set_propval` => `set`), make them chainable. #822